requirements.txt
# ============================================
# Core deep learning stack
# ============================================
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ============================================
# Fast nearest-neighbour search
# ============================================
# Use faiss-gpu when CUDA is available, otherwise faiss-cpu.
# Pinning faiss-cpu by default keeps installation portable.
faiss-cpu>=1.7.4

# ============================================
# Visualisation
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0

# ============================================
# Self-supervised encoders
# ============================================
timm>=0.9.0

# ============================================
# Experiment tracking + utilities
# ============================================
wandb>=0.15.0
tqdm>=4.65.0
pyyaml>=6.0
Pillow>=10.0.0
